almost finished!
tests have been done, we almost made sure of every feature of the project.
only tests on the notification system and the watchlist are all that is left.
notification system is nearly 100% done and ready from the front end side but only tests with some real backend notification system are needed.
watchlist has some simple issues and it needs some editing.